refactor: moved conponents from page directory to all conponents directory
