+++ app/components/FallenList/style.module.scss
@@ -1,0 +1,15 @@
+@use '@/styles/_vars.scss' as *;
+
+.cardBackground {
+    background-color: $elementGradient;
+    border-bottom: 1px solid $lightblueBorder;
+    border-left: 1px solid $lightblueBorder;
+    border-right: 1px solid $lightblueBorder;
+    border-radius: 0px 0px 15px 15px;
+    padding: 10px;
+    width: 250px;
+
+    &:hover {
+        cursor: pointer;
+    }
+}
